requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
aiohttp==3.8.5
PyMySQL=1.1.0
SQLAlchemy=2.0.21
greenlet=3.0.0
aiomysql=0.2.0